Extractor/requirements.txt
tabula-py==2.10.0
pdfplumber==0.11.7
pandas>=2.0.0
openpyxl>=3.1.0
mysql-connector-python>=9.0.0